bake-decouple: move production pipeline into bake, extract bake_types seam
