+++ src/bundler/lib.rs
@@ -305,7 +305,7 @@ pub mod options {
     }
 
     /// Which graph an output belongs to.
-    /// Re-export of the canonical def in `crate::bake_types` (bundle_v2.rs).
+    /// Re-export of the canonical def in `crate::bake_types`.
     pub use crate::bake_types::Side;
 
     pub use crate::bake_types::Framework;
@@ -318,12 +318,11 @@ pub mod options {
 pub use cache::RuntimeTranspilerCacheExt;
 
 // ──────────────────────────────────────────────────────────────────────────
-// Re-export the canonical `bake_types` defs from
-// `bundle_v2` so there is exactly ONE nominal `Side`/`Graph`/`Framework` etc.
-// across the crate (the previous inline copy here diverged and produced
-// "expected `bake_types::Graph`, found `bake_types::Graph`" errors).
+// TYPE_ONLY seam module shared with `bun_runtime::bake` — the single nominal
+// `Side`/`Graph`/`Framework` etc. across the crate. Part of the
+// `DevServerHandle` seam below (its vtable slots name `Graph`/`CacheEntry`).
 // ──────────────────────────────────────────────────────────────────────────
-pub use bundle_v2::bake_types;
+pub mod bake_types;
 
 // ──────────────────────────────────────────────────────────────────────────
 // Re-export the canonical `dispatch` module from
